Now Delay.Listened can also pass a value like Delay.Timeout
The callback for Delay.Listened is changed, one more arg will accept the passed value from the previous Delay on the list.

Also, The test cases for this feature is added. The test framework still seems buggy, so a re-write is applied.

More point-free util is added. Currying seems to be something unavoidable.