Update pretty much things and also break much
It turns out that all of this work is another failure.

The main mistake is that method `ReactionCollector#collect` method does not know about user interface. However, it applies a lot of things to a implied user interface, which is called `@ui` in the collector instance. It is true that the process about user interface will not be executed until `ReactionCollector#reaction` is called and a usable user interface is passed in. But the refactor will break too much things. It will not be easy any more. I do not want to modify the comments, either.

So goodbye. I will not give up, and will not erase this version (it will live forever in this git repo). Time to move on.